+++ release-notes.md
@@ -6,12 +6,15 @@ For players:
 * The console is now simpler and easier to read.
 * The console now adjusts its colors when you have a light terminal background.
 * SMAPI now detects mods which may impact game stability and shows a warning in the console.
-* Updated compatibility list.
+* SMAPI now alerts you in the console when one of your mods has a new version.
 * Renamed installer folder from `SMAPI 2.0` to `SMAPI 2.0 installer` to avoid confusion.
+* Updated compatibility list.
+* Fixed update check errors on Linux/Mac.
 
 For mod developers:
 * Added new APIs to edit, inject, and reload XNB assets loaded by the game at any time.  
   <small>_This let mods do anything previously only possible with XNB mods, plus enables new mod scenarios (e.g. seasonal textures, NPC clothing that depend on the weather or location, etc)._</small>
+* Added new manifest fields to enable automatic update checks.
 * Added new input events.  
   <small>_The new `InputEvents` combine keyboard + mouse + controller input into one event for easy handling, add metadata like the cursor position and grab tile to support click handling, and add an option to suppress input from the game to enable new scenarios like action highjacking and UI overlays._</small>
 * Added support for optional dependencies.